0.0.87: command interface merged
