+++ doc/EXAMPLE.md
@@ -8,17 +8,17 @@ Original picture:
 
 Assuming you are located at the root of the repo, you can generate a 8 colors palette for the example picture `lena.jpg`:
 
-	$ ./pixam-palette doc/examples/lena.jpg 8 doc/examples/palette-8-lena.png
+	$ ./pixam palette doc/examples/lena.jpg 8 doc/examples/palette-8-lena.png
 
 You can now dither this same picture, using for example the `pattern-bubble8.png` pattern:
 
-	$ ./pixam-dither doc/examples/lena.jpg doc/examples/palette-8-lena.png patterns/pattern-bubble8.png doc/examples/lena-bubble.png
+	$ ./pixam dither doc/examples/lena.jpg doc/examples/palette-8-lena.png patterns/pattern-bubble8.png doc/examples/lena-bubble.png
 
 ![Lena Bubble8](examples/lena-bubble.png)
 
 You can now dither this same picture, using for example the `pattern-knut-odith.png` pattern:
 
-	$ ./pixam-dither doc/examples/lena.jpg doc/examples/palette-8-lena.png patterns/pattern-knut-odith.png doc/examples/lena-knut-odith.png
+	$ ./pixam dither doc/examples/lena.jpg doc/examples/palette-8-lena.png patterns/pattern-knut-odith.png doc/examples/lena-knut-odith.png
 
 ![Lena D.E.Knut odith](examples/lena-knut-odith.png)
